also the other places where we use the did, add tests

+++ packages/http-client/src/remote-admin-api.ts
@@ -59,8 +59,10 @@ export class RemoteAdminApi implements AdminApi {
 
   async getIndexedModels(): Promise<Array<StreamID>> {
     const code = await this.generateCode()
+    const did = this._getDidFn()
+    if (!did) throw new MissingDIDError()
     const response = await this._fetchJson(this.getModelsUrl(), {
-      headers: { Authorization: `Basic ${await this.buildJWS(this._getDidFn(), code)}` },
+      headers: { Authorization: `Basic ${await this.buildJWS(did, code)}` },
     })
     return response.models.map((modelStreamIDString: string) => {
       return StreamID.fromString(modelStreamIDString)
@@ -69,9 +71,11 @@ export class RemoteAdminApi implements AdminApi {
 
   async stopIndexingModels(modelsIDs: Array<StreamID>): Promise<void> {
     const code = await this.generateCode()
+    const did = this._getDidFn()
+    if (!did) throw new MissingDIDError()
     await this._fetchJson(this.getModelsUrl(), {
       method: 'delete',
-      body: { jws: await this.buildJWS(this._getDidFn(), code, modelsIDs) },
+      body: { jws: await this.buildJWS(did, code, modelsIDs) },
     })
   }
 }
